Add support for Linux in HTTP client (close #5)
PR #49
