+++ performance/logs.txt
@@ -4,3 +4,4 @@
 {"desktop_context":{"data":{"deviceManufacturer":"Apple Inc.","deviceModel":"MacBookPro17,1","deviceProcessorCount":8,"osIs64Bit":true,"osServicePack":"","osType":"macOS","osVersion":"12.2.0"},"schema":"iglu:com.snowplowanalytics.snowplow/desktop_context/jsonschema/1-0-0"},"results":{"mocked_emitter_and_mocked_session":5.217072375,"mocked_emitter_and_real_session":5.673155916,"mute_emitter_and_mocked_session":36.860906125,"mute_emitter_and_real_session":33.030649,"num_operations":10000,"num_threads":5},"timestamp":1645005862283,"tracker_version":"cpp-0.1.0"}
 {"desktop_context":{"data":{"deviceManufacturer":"Apple Inc.","deviceModel":"MacBookPro17,1","deviceProcessorCount":8,"osIs64Bit":true,"osServicePack":"","osType":"macOS","osVersion":"12.2.0"},"schema":"iglu:com.snowplowanalytics.snowplow/desktop_context/jsonschema/1-0-0"},"results":{"mocked_emitter_and_mocked_session":5.169725125,"mocked_emitter_and_real_session":5.426304083,"mute_emitter_and_mocked_session":27.828702709,"mute_emitter_and_real_session":32.127459,"num_operations":10000,"num_threads":5},"timestamp":1645006143834,"tracker_version":"cpp-0.1.0"}
 {"desktop_context":{"data":{"deviceManufacturer":"Apple Inc.","deviceModel":"MacBookPro17,1","deviceProcessorCount":8,"osIs64Bit":true,"osServicePack":"","osType":"macOS","osVersion":"12.2.0"},"schema":"iglu:com.snowplowanalytics.snowplow/desktop_context/jsonschema/1-0-0"},"results":{"mocked_emitter_and_mocked_session":5.742478167,"mocked_emitter_and_real_session":5.050215375,"mute_emitter_and_mocked_session":25.8688495,"mute_emitter_and_real_session":19.645383417,"num_operations":10000,"num_threads":5},"timestamp":1645041474207,"tracker_version":"cpp-0.1.0"}
+{"desktop_context":{"data":{"deviceManufacturer":"Apple Inc.","deviceModel":"MacBookPro17,1","deviceProcessorCount":8,"osIs64Bit":true,"osServicePack":"","osType":"macOS","osVersion":"12.2.0"},"schema":"iglu:com.snowplowanalytics.snowplow/desktop_context/jsonschema/1-0-0"},"results":{"mocked_emitter_and_mocked_session":6.493122875,"mocked_emitter_and_real_session":6.679922959,"mute_emitter_and_mocked_session":18.915125958,"mute_emitter_and_real_session":17.245701792,"num_operations":10000,"num_threads":5},"timestamp":1648794323772,"tracker_version":"cpp-0.2.0"}
