+++ snowplow-cpp-tracker-example.vcxproj
@@ -103,8 +103,8 @@
     <ClCompile Include="src\client_session.cpp" />
     <ClCompile Include="src\cracked_url.cpp" />
     <ClCompile Include="src\emitter.cpp" />
-    <ClCompile Include="src\http_client.cpp" />
-    <ClCompile Include="src\http_request_result.cpp" />
+    <ClCompile Include="src\http\http_client_windows.cpp" />
+    <ClCompile Include="src\http\http_request_result.cpp" />
     <ClCompile Include="src\payload.cpp" />
     <ClCompile Include="src\self_describing_json.cpp" />
     <ClCompile Include="src\storage.cpp" />
@@ -120,8 +120,9 @@
     <ClInclude Include="src\constants.hpp" />
     <ClInclude Include="src\cracked_url.hpp" />
     <ClInclude Include="src\emitter.hpp" />
-    <ClInclude Include="src\http_client.hpp" />
-    <ClInclude Include="src\http_request_result.hpp" />
+    <ClInclude Include="src\http\http_client.hpp" />
+    <ClInclude Include="src\http\http_client_windows.hpp" />
+    <ClInclude Include="src\http\http_request_result.hpp" />
     <ClInclude Include="src\payload.hpp" />
     <ClInclude Include="src\self_describing_json.hpp" />
     <ClInclude Include="src\storage.hpp" />
