+++ snowplow-cpp-tracker.vcxproj
@@ -107,19 +107,20 @@
     <ClCompile Include="src\client_session.cpp" />
     <ClCompile Include="src\cracked_url.cpp" />
     <ClCompile Include="src\emitter.cpp" />
-    <ClCompile Include="src\http_client.cpp" />
-    <ClCompile Include="src\http_request_result.cpp" />
+    <ClCompile Include="src\http\http_client_windows.cpp" />
+    <ClCompile Include="src\http\http_request_result.cpp" />
     <ClCompile Include="src\payload.cpp" />
     <ClCompile Include="src\self_describing_json.cpp" />
     <ClCompile Include="src\storage.cpp" />
     <ClCompile Include="src\subject.cpp" />
     <ClCompile Include="src\tracker.cpp" />
     <ClCompile Include="src\utils.cpp" />
+    <ClCompile Include="test\http\test_http_client.cpp" />
     <ClCompile Include="test\client_session_test.cpp" />
     <ClCompile Include="test\cracked_url_test.cpp" />
     <ClCompile Include="test\emitter_test.cpp" />
-    <ClCompile Include="test\http_client_test.cpp" />
-    <ClCompile Include="test\http_request_result_test.cpp" />
+    <ClCompile Include="test\http\http_client_test.cpp" />
+    <ClCompile Include="test\http\http_request_result_test.cpp" />
     <ClCompile Include="test\main.cpp" />
     <ClCompile Include="test\payload_test.cpp" />
     <ClCompile Include="test\self_describing_json_test.cpp" />
@@ -136,14 +137,16 @@
     <ClInclude Include="src\constants.hpp" />
     <ClInclude Include="src\cracked_url.hpp" />
     <ClInclude Include="src\emitter.hpp" />
-    <ClInclude Include="src\http_client.hpp" />
-    <ClInclude Include="src\http_request_result.hpp" />
+    <ClInclude Include="src\http\http_client.hpp" />
+    <ClInclude Include="src\http\http_client_windows.hpp" />
+    <ClInclude Include="src\http\http_request_result.hpp" />
     <ClInclude Include="src\payload.hpp" />
     <ClInclude Include="src\self_describing_json.hpp" />
     <ClInclude Include="src\storage.hpp" />
     <ClInclude Include="src\subject.hpp" />
     <ClInclude Include="src\tracker.hpp" />
     <ClInclude Include="src\utils.hpp" />
+    <ClInclude Include="test\http\test_http_client.hpp" />
     <ClInclude Include="test\catch.hpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
