+++ test/http/test_http_client.hpp
@@ -11,41 +11,12 @@ software distributed under the Apache License Version 2.0 is distributed on an
 See the Apache License Version 2.0 for the specific language governing permissions and limitations there under.
 */
 
-#ifndef HTTP_CLIENT_H
-#define HTTP_CLIENT_H
+#ifndef HTTP_CLIENT_TEST_H
+#define HTTP_CLIENT_TEST_H
 
-#include <map>
-#include <string>
-#include <regex>
-#include "constants.hpp"
-#include "cracked_url.hpp"
-#include "http_request_result.hpp"
+#include "../../src/http/http_client.hpp"
 
-#if defined(SNOWPLOW_TEST_SUITE)
-
-#include <iostream>
-#include <thread>
 #include <mutex>
-
-#elif defined(WIN32) || defined(_WIN32) || defined(__WIN32) && !defined(__CYGWIN__)
-
-#include <windows.h>
-#include <WinInet.h>
-#include <tchar.h>
-
-#pragma comment (lib, "wininet.lib")
-
-#elif defined(__APPLE__)
-
-#include <iostream>
-#include <sstream>
-#include <CoreFoundation/CoreFoundation.h>
-#include <CFNetwork/CFNetwork.h>
-#include <CFNetwork/CFHTTPStream.h>
-
-#pragma GCC diagnostic ignored "-Wdeprecated-declarations"
-
-#endif
 
 using std::string;
 using std::list;
@@ -55,15 +26,12 @@ namespace snowplow {
 /**
  * @brief HTTP client for making requests to Snowplow Collector. To be used internally within tracker only.
  */
-class HttpClient {
+class TestHttpClient : public HttpClient {
 public:
-  enum RequestMethod { POST, GET };
-  
-  static const string TRACKER_AGENT;
-  static HttpRequestResult http_post(const CrackedUrl url, const string & post_data, list<int> row_ids, bool oversize);
-  static HttpRequestResult http_get(const CrackedUrl url, const string & query_string, list<int> row_ids, bool oversize);
+  ~TestHttpClient() {}
 
-#if defined(SNOWPLOW_TEST_SUITE)
+  static const string TRACKER_AGENT;
+
   struct Request {
     Request(){};
     RequestMethod method;
@@ -80,10 +48,9 @@ public:
   static void set_http_response_code(int http_response_code);
   static list<Request> get_requests_list();
   static void reset();
-#endif
 
-private:
-  static HttpRequestResult http_request(const RequestMethod method, const CrackedUrl url, const string & query_string, const string & post_data, list<int> row_ids, bool oversize);
+protected:
+  HttpRequestResult http_request(const RequestMethod method, const CrackedUrl url, const string & query_string, const string & post_data, list<int> row_ids, bool oversize);
 };
 }
